requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
python-dotenv>=1.0,<2.0
streamlit>=1.38,<2.0
requests>=2.32,<3.0
chromadb>=0.5,<1.0
sentence-transformers>=3.0,<4.0
Pillow>=10.0,<12.0
mcp>=1.0,<2.0
